disktec-main/requirements.txt
asgiref==3.5.2
black==22.6.0
boto3==1.24.43
botocore==1.27.43
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
distlib==0.3.4
Django==4.0.6
django-environ==0.9.0
django-storages==1.12.3
Faker==13.14.0
filelock==3.7.1
gunicorn==20.1.0
identify==2.5.1
idna==3.3
jmespath==1.0.1
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.19.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.15.1
whitenoise==6.2.0
